~~ Addon 2: Preserve sources order during incremental compilation
The
`IncrementalJsFirKlibCompilerWithScopeExpansionRunnerTestGenerated
.testCompanionObjectInheritedMemberChanged`, and
`IncrementalJsFirKlibCompilerWithScopeExpansionRunnerTestGenerated
.testCompanionObjectMemberChanged`
tests fail because the checksums of
the `manifest` files don't match
after the full recompilation.
It happens, because some irrelevant
IrVariables' names change from,
for example, `tmp2_safe_receiver`
to `tmp0_safe_receiver`. That is
the compiler receives the input
files
in arbitrary order (because
when we calculate the dirty files
that must be recalculated after a
step, we get them in arbitrary
order.

This change removes the flaky
behavior by preserving the order.