~~ Addon 1: Reset temporary vars index per FirFile
The
`IncrementalJsFirKlibCompilerWithScopeExpansionRunnerTestGenerated
.testCompanionObjectInheritedMemberChanged`, and
`IncrementalJsFirKlibCompilerWithScopeExpansionRunnerTestGenerated
.testCompanionObjectMemberChanged`
tests fail because the checksums of
the `manifest` files don't match
after the full recompilation.
It happens, because some irrelevant
IrVariables' names change from,
for example, `tmp2_safe_receiver`
to `tmp0_safe_receiver`. That is
the compiler receives the input
files
in arbitrary order (because
when we calculate the dirty files
that must be recalculated after a
step, we get them in arbitrary
order.

This change removes the flaky
behavior by ensuring the temporary
names stay the same independently
of "when" the file is recompiled.