Fix generator test to java reflection Argument
The old test was testing a feature not used in practice; namely that the
Argument annotation could be specified on an interface property, without
a java backing field. All actual uses of the annotation are on
properties with a java backing field. This fix changes the interface to
be a class instead, in line with actual use cases.